Create common library for services
