+++ mercury-framework/customer-credit-rating/customer-credit-rating-service/pom.xml
@@ -16,6 +16,11 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
+            <groupId>com.redhat.mercury</groupId>
+            <artifactId>mercury-common-service</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-junit5</artifactId>
             <scope>test</scope>
