--- mercury-framework/deploy/examples/00_customer-credit-rating.yaml
@@ -1,10 +1,0 @@
-apiVersion: mercury.redhat.io/v1alpha1
-kind: ServiceDomainBinding
-metadata:
-  name: customer-credit-rating-bindings
-  labels:
-    service-domain: customer-credit-rating
-spec:
-  exposedScopes:
-    - scopeRef: customercreditrating
-      action: query
