--- mercury-framework/deploy/examples/00_party-routing.yaml
@@ -1,15 +1,0 @@
-apiVersion: mercury.redhat.io/v1alpha1
-kind: ServiceDomainBinding
-metadata:
-  name: party-routing-profile-bindings
-  labels:
-    service-domain: party-routing-profile
-spec:
-  exposedScopes:
-    - scopeRef: partyroutingprofile
-      action: query
-  subscriptions:
-    - serviceDomain: customer-offer
-      events:
-        - org.bian.customeroffer.initiated
-        - org.bian.customeroffer.completed
