+++ mercury-framework/mercury-common-service/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+quarkus.grpc.server.port=10001
+quarkus.grpc.clients.outboundBindingService.port=10100
+quarkus.grpc.clients.outboundBindingService.host=0.0.0.0
