+++ mercury-framework/mercury-common-service/src/test/resources/application.properties
@@ -1,0 +1,1 @@
+
